refactor(snowman): extract collision handling (R3.11)
